refactor: move helpers under `utils` module
